+++ Classes/Domain/Model/Transfer.php
@@ -110,6 +110,11 @@ class Transfer extends AbstractEntity
         return $this->crdate;
     }
 
+    public function setCrdate(int $crdate): void
+    {
+        $this->crdate = $crdate;
+    }
+
     public function getStepUid(): int
     {
         return $this->stepUid;